requirements.txt
Flask==2.3.2
pandas==2.0.3
Pillow==10.0.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
requests==2.31.0
celery==5.3.4